✨ refactor hooks to use consistent naming convention and improve code organization
