+++ src/hooks/useFetchDocument.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react"
 import { db } from "../lib/firebase/config"
 import { doc, getDoc } from 'firebase/firestore'
 
-export const userFetchDocument = (docCollection, id) => {
+export const useFetchDocument = (docCollection, id) => {
     const [document, setDocument] = useState(null)
     const [error, setError] = useState(null)
     const [loading, setLoading] = useState(null)
