+++ src/hooks/useFetchDocuments.jsx
@@ -8,7 +8,7 @@ import {
   where,
 } from "firebase/firestore";
 
-export const userFetchDocuments = (docCollection, search = null, uid = null) => {
+export const useFetchDocuments = (docCollection, search = null, uid = null) => {
   const [documents, setDocuments] = useState(null);
   const [error, setError] = useState(null);
   const [loading, setLoading] = useState(null);
