+++ src/hooks/useQuery.jsx
@@ -1,7 +1,7 @@
 import { useLocation } from "react-router-dom";
 import { useMemo } from "react";
 
-export function userQuery() {
+export function useQuery() {
   const { search } = useLocation();
 
   return useMemo(() => new URLSearchParams(search), [search]);
