+++ src/pages/Dashboard/Dashboard.jsx
@@ -3,16 +3,16 @@ import styles from "./Dashboard.module.css";
 import { Link } from "react-router-dom";
 
 import { useAuthValue } from "../../context/AuthContext";
-import { userFetchDocuments } from "../../hooks/userFetchDocuments";
-import { userDeleteDocument } from "../../hooks/userDeleteDocument";
+import { useFetchDocuments } from "../../hooks/useFetchDocuments";
+import { useDeleteDocument } from "../../hooks/useDeleteDocument";
 
 const Dashboard = () => {
   const { user } = useAuthValue();
   const uid = user.uid;
 
-  const { documents: posts } = userFetchDocuments("posts", null, uid);
+  const { documents: posts } = useFetchDocuments("posts", null, uid);
 
-  const { deleteDocument } = userDeleteDocument("posts");
+  const { deleteDocument } = useDeleteDocument("posts");
 
   console.log(uid);
   console.log(posts);
