+++ src/pages/Login/Login.jsx
@@ -1,14 +1,14 @@
 import styles from "./Login.module.css"
 
 import { useEffect, useState } from "react"
-import { userAuthentication } from "../../hooks/userAuthentication"
+import { useAuthentication } from "../../hooks/useAuthentication"
 
 const Login = () => {
   const [email, setEmail] = useState("")
   const [password, setPassword] = useState("")
   const [error, setError] = useState("")
 
-  const { login, error: authError, loading } = userAuthentication()
+  const { login, error: authError, loading } = useAuthentication()
 
   const handleSubmit = async (e) => {
     e.preventDefault()
