+++ src/pages/Register/Register.jsx
@@ -1,7 +1,7 @@
 import styles from "./Register.module.css"
 
 import { useEffect, useState } from "react"
-import { userAuthentication } from "../../hooks/userAuthentication"
+import { useAuthentication } from "../../hooks/useAuthentication"
 
 const Register = () => {
   const [displayName, setDisplayName] = useState("")
@@ -10,7 +10,7 @@ const Register = () => {
   const [confirmPassword, setConfirmPassword] = useState("")
   const [error, setError] = useState("")
 
-  const { createUser, error: authError, loading } = userAuthentication()
+  const { createUser, error: authError, loading } = useAuthentication()
 
   const handleSubmit = async (e) => {
     e.preventDefault()
